moved files and added helpers.ts
