+++ src/setup-server.ts
@@ -2,7 +2,7 @@ import { ApolloServer } from "@apollo/server";
 import { startStandaloneServer } from "@apollo/server/standalone";
 import { typeDefs } from "./graphql/schema";
 import { resolvers } from "./graphql/resolvers";
-import { formatError } from "./graphql/error-handlers";
+import { formatError } from "./graphql/helpers/error-handlers";
 
 export let server: ApolloServer;
 export let url: string;
